+++ packages/@uppy/companion-client/src/index.js
@@ -6,10 +6,12 @@
 
 const RequestClient = require('./RequestClient')
 const Provider = require('./Provider')
+const SearchProvider = require('./SearchProvider')
 const Socket = require('./Socket')
 
 module.exports = {
   RequestClient,
   Provider,
+  SearchProvider,
   Socket
 }
