+++ packages/@uppy/companion/src/server/helpers/request.js
@@ -157,7 +157,7 @@ class HttpsAgent extends https.Agent {
  *
  * @param {string} url
  * @param {boolean=} blockLocalIPs
- * @return {Promise}
+ * @return {Promise<{type: string, size: number}>}
  */
 exports.getURLMeta = (url, blockLocalIPs = false) => {
   return new Promise((resolve, reject) => {
