+++ packages/@uppy/provider-views/src/CloseWrapper.js
@@ -1,0 +1,11 @@
+const { h, Component } = require('preact')
+
+module.exports = class CloseWrapper extends Component {
+  componentWillUnmount () {
+    this.props.onUnmount()
+  }
+
+  render () {
+    return this.props.children[0]
+  }
+}
