+++ packages/@uppy/provider-views/src/ProviderView/AuthView.js
