+++ packages/@uppy/provider-views/src/SearchProviderView/index.js
@@ -1,0 +1,1 @@
+module.exports = require('./SearchProviderView')
