+++ packages/@uppy/provider-views/src/style.scss
@@ -4,6 +4,7 @@
 @import './style/uppy-ProviderBrowser-viewType--grid';
 @import './style/uppy-ProviderBrowser-viewType--list';
 @import './style/uppy-ProviderBrowserItem-fakeCheckbox';
+@import './style/uppy-SearchProvider-input.scss';
 
 .uppy-DashboardContent-panelBody {
   display: flex;
