+++ packages/@uppy/provider-views/src/style/uppy-SearchProvider-input.scss
@@ -1,0 +1,32 @@
+
+.uppy-SearchProvider {
+  width: 100%;
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  flex: 1;
+
+  [data-uppy-theme="dark"] & {
+    background-color: $gray-900;
+  }
+}
+
+.uppy-SearchProvider-input {
+  width: 90%;
+  max-width: 650px;
+  margin-bottom: 15px;
+
+  .uppy-size--md & {
+    margin-bottom: 20px;
+  }
+}
+
+.uppy-SearchProvider-searchButton {
+  padding: 13px 25px;
+
+  .uppy-size--md & {
+    padding: 13px 30px;
+  }
+}
