+++ packages/@uppy/unsplash/types/index.test-d.ts
@@ -1,0 +1,2 @@
+import Unsplash = require('../')
+// TODO implement
