+++ packages/@uppy/utils/src/truncateString.js
