+++ packages/@uppy/utils/src/truncateString.test.js
