+++ packages/uppy/index.js
@@ -26,6 +26,7 @@ exports.GoogleDrive = require('@uppy/google-drive')
 exports.Instagram = require('@uppy/instagram')
 exports.OneDrive = require('@uppy/onedrive')
 exports.Facebook = require('@uppy/facebook')
+exports.Unsplash = require('@uppy/unsplash')
 exports.Url = require('@uppy/url')
 exports.Webcam = require('@uppy/webcam')
 exports.ScreenCapture = require('@uppy/screen-capture')
